+++ components/fxa-client/src/http_client.rs
@@ -5,6 +5,7 @@
 use crate::{config::Config, errors::*};
 use serde_derive::*;
 use serde_json::json;
+use std::collections::HashMap;
 use viaduct::{header_names, status_codes, Request, Response};
 
 #[cfg(feature = "browserid")]
@@ -30,6 +31,28 @@ pub trait FxAClient {
         profile_access_token: &str,
         etag: Option<String>,
     ) -> Result<Option<ResponseAndETag<ProfileResponse>>>;
+    fn pending_commands(
+        &self,
+        config: &Config,
+        refresh_token: &str,
+        index: u64,
+        limit: Option<u64>,
+    ) -> Result<PendingCommandsResponse>;
+    fn invoke_command(
+        &self,
+        config: &Config,
+        refresh_token: &str,
+        command: &str,
+        target: &str,
+        payload: &serde_json::Value,
+    ) -> Result<()>;
+    fn devices(&self, config: &Config, refresh_token: &str) -> Result<Vec<GetDeviceResponse>>;
+    fn update_device(
+        &self,
+        config: &Config,
+        refresh_token: &str,
+        update: DeviceUpdateRequest<'_>,
+    ) -> Result<UpdateDeviceResponse>;
 }
 
 pub struct Client;
@@ -37,14 +60,12 @@ impl FxAClient for Client {
     fn profile(
         &self,
         config: &Config,
-        profile_access_token: &str,
+        access_token: &str,
         etag: Option<String>,
     ) -> Result<Option<ResponseAndETag<ProfileResponse>>> {
         let url = config.userinfo_endpoint()?;
-        let mut request = Request::get(url).header(
-            header_names::AUTHORIZATION,
-            format!("Bearer {}", profile_access_token),
-        )?;
+        let mut request =
+            Request::get(url).header(header_names::AUTHORIZATION, bearer_token(access_token))?;
         if let Some(etag) = etag {
             request = request.header(header_names::IF_NONE_MATCH, format!("\"{}\"", etag))?;
         }
@@ -99,21 +120,80 @@ impl FxAClient for Client {
         Self::make_request(Request::post(url).json(&body))?;
         Ok(())
     }
+
+    fn pending_commands(
+        &self,
+        config: &Config,
+        refresh_token: &str,
+        index: u64,
+        limit: Option<u64>,
+    ) -> Result<PendingCommandsResponse> {
+        let url = config.auth_url_path("v1/account/device/commands")?;
+        let mut request = Request::get(url)
+            .header(header_names::AUTHORIZATION, bearer_token(refresh_token))?
+            .query(&[("index", &index.to_string())]);
+        if let Some(limit) = limit {
+            request = request.query(&[("limit", &limit.to_string())])
+        }
+        Ok(Self::make_request(request)?.json()?)
+    }
+
+    fn invoke_command(
+        &self,
+        config: &Config,
+        refresh_token: &str,
+        command: &str,
+        target: &str,
+        payload: &serde_json::Value,
+    ) -> Result<()> {
+        let body = json!({
+            "command": command,
+            "target": target,
+            "payload": payload
+        });
+        let url = config.auth_url_path("v1/account/devices/invoke_command")?;
+        let request = Request::post(url)
+            .header(header_names::AUTHORIZATION, bearer_token(refresh_token))?
+            .header(header_names::CONTENT_TYPE, "application/json")?
+            .body(body.to_string());
+        Self::make_request(request)?;
+        Ok(())
+    }
+
+    fn devices(&self, config: &Config, refresh_token: &str) -> Result<Vec<GetDeviceResponse>> {
+        let url = config.auth_url_path("v1/account/devices")?;
+        let request =
+            Request::get(url).header(header_names::AUTHORIZATION, bearer_token(refresh_token))?;
+        Ok(Self::make_request(request)?.json()?)
+    }
+
+    fn update_device(
+        &self,
+        config: &Config,
+        refresh_token: &str,
+        update: DeviceUpdateRequest<'_>,
+    ) -> Result<UpdateDeviceResponse> {
+        let url = config.auth_url_path("v1/account/device")?;
+        let request = Request::post(url)
+            .header(header_names::AUTHORIZATION, bearer_token(refresh_token))?
+            .header(header_names::CONTENT_TYPE, "application/json")?
+            .body(serde_json::to_string(&update)?);
+        Ok(Self::make_request(request)?.json()?)
+    }
 }
 
 impl Client {
     pub fn new() -> Self {
         Self {}
     }
+
     fn make_oauth_token_request(
         &self,
         config: &Config,
         body: serde_json::Value,
     ) -> Result<OAuthTokenResponse> {
         let url = config.token_endpoint()?;
-        Self::make_request(Request::post(url).json(&body))?
-            .json()
-            .map_err(Into::into)
+        Ok(Self::make_request(Request::post(url).json(&body))?.json()?)
     }
 
     fn make_request(request: Request) -> Result<Response> {
@@ -137,10 +217,181 @@ impl Client {
     }
 }
 
+fn bearer_token(token: &str) -> String {
+    format!("Bearer {}", token)
+}
+
 #[derive(Clone)]
 pub struct ResponseAndETag<T> {
     pub response: T,
     pub etag: Option<String>,
+}
+
+#[derive(Deserialize)]
+pub struct PendingCommandsResponse {
+    pub index: u64,
+    pub last: Option<bool>,
+    pub messages: Vec<PendingCommand>,
+}
+
+#[derive(Deserialize)]
+pub struct PendingCommand {
+    pub index: u64,
+    pub data: CommandData,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct CommandData {
+    pub command: String,
+    pub payload: serde_json::Value, // Need https://github.com/serde-rs/serde/issues/912 to make payload an enum instead.
+    pub sender: Option<String>,
+}
+
+#[derive(Clone, Deserialize, Serialize)]
+pub struct PushSubscription {
+    #[serde(rename = "pushCallback")]
+    pub endpoint: String,
+    #[serde(rename = "pushPublicKey")]
+    pub public_key: String,
+    #[serde(rename = "pushAuthKey")]
+    pub auth_key: String,
+}
+
+/// We use the double Option pattern in this struct.
+/// The outer option represents the existence of the field
+/// and the inner option its value or null.
+/// TL;DR:
+/// `None`: the field will not be present in the JSON body.
+/// `Some(None)`: the field will have a `null` value.
+/// `Some(Some(T))`: the field will have the serialized value of T.
+#[derive(Serialize)]
+#[allow(clippy::option_option)]
+pub struct DeviceUpdateRequest<'a> {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(rename = "name")]
+    display_name: Option<Option<&'a str>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(rename = "type")]
+    device_type: Option<Option<&'a DeviceType>>,
+    #[serde(flatten)]
+    push_subscription: Option<&'a PushSubscription>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(rename = "availableCommands")]
+    available_commands: Option<Option<&'a HashMap<String, String>>>,
+}
+
+#[derive(Clone, Serialize, Deserialize)]
+pub enum DeviceType {
+    #[serde(rename = "desktop")]
+    Desktop,
+    #[serde(rename = "mobile")]
+    Mobile,
+    #[serde(other)]
+    #[serde(skip_serializing)] // Don't you dare trying.
+    Unknown,
+}
+
+#[allow(clippy::option_option)]
+pub struct DeviceUpdateRequestBuilder<'a> {
+    device_type: Option<Option<&'a DeviceType>>,
+    display_name: Option<Option<&'a str>>,
+    push_subscription: Option<&'a PushSubscription>,
+    available_commands: Option<Option<&'a HashMap<String, String>>>,
+}
+
+impl<'a> DeviceUpdateRequestBuilder<'a> {
+    pub fn new() -> Self {
+        Self {
+            device_type: None,
+            display_name: None,
+            push_subscription: None,
+            available_commands: None,
+        }
+    }
+
+    pub fn push_subscription(mut self, push_subscription: &'a PushSubscription) -> Self {
+        self.push_subscription = Some(push_subscription);
+        self
+    }
+
+    pub fn available_commands(mut self, available_commands: &'a HashMap<String, String>) -> Self {
+        self.available_commands = Some(Some(available_commands));
+        self
+    }
+
+    pub fn clear_available_commands(mut self) -> Self {
+        self.available_commands = Some(None);
+        self
+    }
+
+    pub fn display_name(mut self, display_name: &'a str) -> Self {
+        self.display_name = Some(Some(display_name));
+        self
+    }
+
+    pub fn clear_display_name(mut self) -> Self {
+        self.display_name = Some(None);
+        self
+    }
+
+    #[allow(dead_code)]
+    pub fn device_type(mut self, device_type: &'a DeviceType) -> Self {
+        self.device_type = Some(Some(device_type));
+        self
+    }
+
+    pub fn build(self) -> DeviceUpdateRequest<'a> {
+        DeviceUpdateRequest {
+            display_name: self.display_name,
+            device_type: self.device_type,
+            push_subscription: self.push_subscription,
+            available_commands: self.available_commands,
+        }
+    }
+}
+
+#[derive(Clone, Serialize, Deserialize)]
+pub struct DeviceLocation {
+    pub city: Option<String>,
+    pub country: Option<String>,
+    pub state: Option<String>,
+    #[serde(rename = "stateCode")]
+    pub state_code: Option<String>,
+}
+
+#[derive(Clone, Serialize, Deserialize)]
+pub struct GetDeviceResponse {
+    #[serde(flatten)]
+    pub common: DeviceResponseCommon,
+    #[serde(rename = "isCurrentDevice")]
+    pub is_current_device: bool,
+    pub location: DeviceLocation,
+    #[serde(rename = "lastAccessTime")]
+    pub last_access_time: Option<u64>,
+}
+
+impl std::ops::Deref for GetDeviceResponse {
+    type Target = DeviceResponseCommon;
+    fn deref(&self) -> &DeviceResponseCommon {
+        &self.common
+    }
+}
+
+pub type UpdateDeviceResponse = DeviceResponseCommon;
+
+#[derive(Clone, Serialize, Deserialize)]
+pub struct DeviceResponseCommon {
+    pub id: String,
+    #[serde(rename = "name")]
+    pub display_name: String,
+    #[serde(rename = "type")]
+    pub device_type: DeviceType,
+    #[serde(flatten)]
+    pub push_subscription: Option<PushSubscription>,
+    #[serde(rename = "availableCommands")]
+    pub available_commands: HashMap<String, String>,
+    #[serde(rename = "pushEndpointExpired")]
+    pub push_endpoint_expired: bool,
 }
 
 #[derive(Deserialize)]